+++ src/blackboxTests/src/test/kotlin/org/vaccineimpact/api/blackboxTests/BlackboxTestsSuite.kt
@@ -5,10 +5,10 @@ import org.junit.BeforeClass
 import org.junit.runner.RunWith
 import org.junit.runners.Suite
 import org.vaccineimpact.api.blackboxTests.tests.*
-import org.vaccineimpact.api.blackboxTests.tests.BurdenEstimates.BurdenEstimateTests
+import org.vaccineimpact.api.blackboxTests.tests.BurdenEstimates.CreateBurdenEstimateTests
 import org.vaccineimpact.api.blackboxTests.tests.BurdenEstimates.ModelRunParameterTests
 import org.vaccineimpact.api.blackboxTests.tests.BurdenEstimates.RetrieveBurdenEstimateTests
-import org.vaccineimpact.api.blackboxTests.tests.BurdenEstimates.UploadBurdenEstimateTests
+import org.vaccineimpact.api.blackboxTests.tests.BurdenEstimates.PopulateBurdenEstimateTests
 import org.vaccineimpact.api.test_helpers.DatabaseCreationHelper
 
 // Keep these sorted alphabetically, for consistency
@@ -17,6 +17,7 @@ import org.vaccineimpact.api.test_helpers.DatabaseCreationHelper
         AccessLogTests::class,
         AuthenticationTests::class,
         CoverageTests::class,
+        CreateBurdenEstimateTests::class,
         CreateUserTests::class,
         DemographicTests::class,
         DiseaseTests::class,
@@ -26,11 +27,11 @@ import org.vaccineimpact.api.test_helpers.DatabaseCreationHelper
         ModelRunParameterTests::class,
         ModelTests::class,
         PasswordTests::class,
+        PopulateBurdenEstimateTests::class,
         ResponsibilityTests::class,
         RetrieveBurdenEstimateTests::class,
         ScenarioTests::class,
         TouchstoneTests::class,
-        UploadBurdenEstimateTests::class,
         UserTests::class
 )
 class BlackboxTestsSuite
