+++ src/blackboxTests/src/test/kotlin/org/vaccineimpact/api/blackboxTests/tests/BurdenEstimates/PopulateBurdenEstimateTests.kt
@@ -1,6 +1,5 @@
 package org.vaccineimpact.api.blackboxTests.tests.BurdenEstimates
 
-import com.beust.klaxon.json
 import org.assertj.core.api.Assertions
 import org.junit.Test
 import org.vaccineimpact.api.blackboxTests.helpers.*
@@ -10,37 +9,8 @@ import org.vaccineimpact.api.models.permissions.PermissionSet
 import org.vaccineimpact.api.validateSchema.JSONValidator
 import spark.route.HttpMethod
 
-
-class UploadBurdenEstimateTests : BurdenEstimateTests()
+class PopulateBurdenEstimateTests : BurdenEstimateTests()
 {
-    private val createdSetLocation = LocationConstraint(
-            "/modelling-groups/group-1/responsibilities/touchstone-1/scenario-1/estimates/", unknownId = true
-    )
-
-    @Test
-    fun `can create burden estimate set without model run parameter set`()
-    {
-        validate(setUrl, method = HttpMethod.post) withRequestSchema "CreateBurdenEstimateSet" given { db ->
-            setUp(db)
-        } sendingJSON {
-            metadataForCreate()
-        } withPermissions {
-            requiredWritePermissions.plus(PermissionSet("*/can-login"))
-        } andCheckObjectCreation createdSetLocation
-    }
-
-    @Test
-    fun `can create burden estimate set with model run parameter set`()
-    {
-        validate(setUrl, method = HttpMethod.post) withRequestSchema "CreateBurdenEstimateSet" given { db ->
-            setUpWithModelRunParameterSet(db)
-        } sendingJSON {
-            metadataForCreateWithModelRunParameterSet()
-        } withPermissions {
-            requiredWritePermissions.plus(PermissionSet("*/can-login"))
-        } andCheckObjectCreation createdSetLocation
-    }
-
     @Test
     fun `can populate central burden estimate`()
     {
@@ -168,41 +138,6 @@ class UploadBurdenEstimateTests : BurdenEstimateTests()
     }
 
     @Test
-    fun `can create burden estimate via onetime link`()
-    {
-        validate("$setUrl/get_onetime_link/") against "Token" given { db ->
-            setUp(db)
-        } requiringPermissions {
-            requiredWritePermissions
-        } andCheckString { token ->
-            val oneTimeURL = "/onetime_link/$token/"
-            val requestHelper = RequestHelper()
-            val response = requestHelper.post(oneTimeURL, metadataForCreate())
-            createdSetLocation.checkObjectCreation(response)
-
-            val badResponse = requestHelper.get(oneTimeURL)
-            JSONValidator().validateError(badResponse.text, expectedErrorCode = "invalid-token-used")
-        }
-    }
-
-    @Test
-    fun `can create burden estimate via onetime link and redirect`()
-    {
-        validate("$setUrl/get_onetime_link/?redirectUrl=http://localhost/") against "Token" given { db ->
-            setUp(db)
-        } requiringPermissions {
-            requiredWritePermissions
-        } andCheckString { token ->
-            val oneTimeURL = "/onetime_link/$token/"
-            val requestHelper = RequestHelper()
-
-            val response = requestHelper.post(oneTimeURL, metadataForCreate())
-            val resultAsString = response.getResultFromRedirect(checkRedirectTarget = "http://localhost")
-            JSONValidator().validateSuccess(resultAsString)
-        }
-    }
-
-    @Test
     fun `bad CSV headers results in ValidationError in redirect`()
     {
         TestUserHelper.setupTestUser()
@@ -222,22 +157,5 @@ class UploadBurdenEstimateTests : BurdenEstimateTests()
             val resultAsString = response.getResultFromRedirect(checkRedirectTarget = "http://localhost")
             JSONValidator().validateError(resultAsString, expectedErrorCode = "csv-unexpected-header")
         }
-    }
-
-    private fun metadataForCreate() = json {
-        obj("type" to obj(
-                "type" to "central-averaged",
-                "details" to "median"
-        ))
-    }
-
-    private fun metadataForCreateWithModelRunParameterSet() = json {
-        obj(
-                "type" to obj(
-                        "type" to "central-averaged",
-                        "details" to "median"
-                ),
-                "model_run_parameter_set" to 1
-        )
     }
 }
